refactor(cli): split 2582-line cli.go into per-command files
Pure mechanical split — no behavior, signature, or logic changes.
cli.go (528 lines) retains the dispatcher, shared helpers, and buildParams.
Command groups moved to: cmd_arming.go, cmd_plant.go, cmd_scan.go,
cmd_events.go, cmd_status.go, cmd_config.go, cmd_serve.go.

Note: "arm" command lives in cmd_arming.go (not cmd_arm.go) because
Go treats the _arm filename suffix as a GOARCH=arm build constraint.
